finished rewriting ReportManagerImplTest.java using jooq a little refactor in a few places

+++ controller/src/main/java/ru/aleksseii/report/ReportManager.java
@@ -4,6 +4,7 @@ import org.jetbrains.annotations.NotNull;
 import ru.aleksseii.model.Organization;
 import ru.aleksseii.model.Product;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.List;
 import java.util.Map;
@@ -16,7 +17,7 @@ public interface ReportManager {
     @NotNull Set<@NotNull Organization> getOrgsWhichSentProductsGTAmount(
             @NotNull Map<@NotNull Integer, @NotNull Integer> productToAmount);
 
-    @NotNull Map<@NotNull Date, @NotNull Map<@NotNull Product, @NotNull List<@NotNull Long>>> getProductAmountAndSumForPeriod(
+    @NotNull Map<@NotNull Date, @NotNull Map<@NotNull Product, @NotNull List<@NotNull BigDecimal>>> getProductAmountAndSumForPeriod(
             @NotNull Date start,
             @NotNull Date end);
